+++ lib/com/flutter/http/adapter/CallBack.dart
@@ -1,6 +1,7 @@
-import 'ResponseDatas.dart';
 
-typedef NetCallbackFunction<T> = void Function(ResponseDatas<T>?);
+import 'package:flutter_dio_module/lib_dio.dart';
+
+typedef NetCallbackFunction<T> = void Function(ResponseDates<T>?);
 
 ///回调
 class CallBack<T> {