+++ lib/com/flutter/http/bean/BaseBean.dart
@@ -14,7 +14,7 @@ class BaseBean<T> {
   BaseBean.fromJson(Map<String, dynamic> json) {
     if (json['data'] != null && json['data'] != 'null') {
       data =
-          RxDioConfig.intstance.getJsonConvert().fromJsonAsT<T>(json['data']);
+          RxDioConfig.instance.getJsonConvert().fromJsonAsT<T>(json['data']);
     }
     code = json['code'] ?? json['errorCode'];
     message = json['message'] ?? json['errorMsg'];