+++ lib/lib_dio.dart
@@ -4,14 +4,14 @@ export 'com/app,data/Constants.dart';
 export 'com/flutter/http/adapter/CallBack.dart';
 export 'com/flutter/http/adapter/Method.dart';
 export 'com/flutter/http/bean/BaseBean.dart';
-export 'com/flutter/http/interceptorss/HttpLogInterceptor.dart';
 export 'com/flutter/http/utils/MD5Utils.dart';
 export 'com/flutter/http/RxDio.dart';
 export 'com/flutter/http/RxDioConfig.dart';
 export 'com/flutter/http/adapter/ResponseDatas.dart';
 
-export 'package:dio/dio.dart';
-export 'package:dio/src/cancel_token.dart';
-export 'package:json_annotation/json_annotation.dart';
-export 'package:rxdart/rxdart.dart';
 export 'package:flutter_dio_module/generated/json/base/i_json_convert.dart';
+export 'package:flutter_dio_module/generated/json/base/base_convert.g.dart';
+
+export 'package:flutter_dio_module/com/flutter/http/cacheUtils/CacheInterface.dart';
+export 'package:flutter_dio_module/com/flutter/http/cacheUtils/CacheInterceptorInterface.dart';
+export 'package:flutter_dio_module/com/flutter/http/cacheUtils/CacheInterceptor.dart';